Telegram_Bot_Cosmos_Proposals/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
clipboard==0.0.4
cryptography==36.0.1
frozenlist==1.3.0
h11==0.13.0
html5lib==1.1
idna==3.3
iso8601==1.0.2
keyboard==0.13.5
lxml==4.8.0
multidict==6.0.2
multitasking==0.0.10
numpy==1.22.3
outcome==1.1.0
pandas==1.4.1
pandas-datareader==0.10.0
pycparser==2.21
pyOpenSSL==21.0.0
pyperclip==1.8.2
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytube==12.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
telepot==12.7
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.7
webencodings==0.5.1
wsproto==1.0.0
yarl==1.7.2
yfinance==0.1.70
